client: initialize struct conn fields after initiating the connection
Best to get the request out the door first, and then do this work after
it's in flight. Not a huge deal, as on my computer we're talking about 10-30 us
saved, but really might as well. Also here we want to make sure to initialize all
we can in conn_init() before aborting on allocation failure so that stuff is correct
on later callbacks (e.g. it's put on client->conn_list)