include dependencies as submodules and build them in this tree
we need libwebsockets to be build with libglib event loop support,
and it's broken on v4.0-stable (takes 100% CPU). So we need at least
v4.1-stable with a particular set of build options. Easiest to just
use submodules and build it here.

For now also we get rid of the shared objects. Should work out how
to do that in a way that makes sense while building deps in tree